# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CategoriesListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
    type: object
  dto.CategoryAnalytics:
    properties:
      average_expense_amount:
        type: number
      average_per_day:
        type: number
      category_id:
        type: integer
      category_name:
        type: string
      expenses_count:
        type: integer
      largest_expense:
        $ref: '#/definitions/dto.ExpenseResponse'
      period:
        type: string
      smallest_expense:
        $ref: '#/definitions/dto.ExpenseResponse'
      total_amount:
        type: number
    type: object
  dto.CategoryPeriod:
    properties:
      period:
        type: string
    type: object
  dto.CategoryResponse:
    properties:
      created_at:
        description: Description *string   `json:"description"`
        type: string
      expenses_count:
        description: Дополнительная информация
        type: integer
      id:
        type: integer
      name:
        type: string
      total_amount:
        type: number
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      category_name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - category_name
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        example: Validation failed
        type: string
      message:
        example: Email is required
        type: string
    type: object
...
  dto.ExpenseResponse:
    properties:
      amount:
        description: Category     CategoryResponse `json:"category,omitempty"`
        type: number
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Система управления личными финансами
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Finance API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Получение всех категорий пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список категорий
          schema:
            $ref: '#/definitions/dto.CategoriesListResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка категорий
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Создание новой категории расходов для пользователя
      parameters:
      - description: Данные для создания категории
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно создана
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой категории
      tags:
      - Categories
  /categories/{category_id}:
    delete:
      consumes:
      - application/json
      description: Удаление категории и всех связанных с ней расходов
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Категория успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID категории
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление категории
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Получение информации о конкретной категории пользователя
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о категории
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Неверный ID категории
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Категория не найдена
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение категории по ID
      tags:
      - Categories
  /categories/{category_id}/analytics:
    get:
      consumes:
      - application/json
      description: Получение детальной аналитики расходов по конкретной категории
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: Период для анализа
        in: body
        name: period
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryPeriod'
      produces:
      - application/json
      responses:
        "200":
          description: Аналитика по категории
          schema:
            $ref: '#/definitions/dto.CategoryAnalytics'
        "400":
          description: Неверный ID категории или период
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение аналитики по категории
      tags:
      - Categories
...
  /categories/top:
    get:
      consumes:
      - application/json
      description: Получение списка категорий, отсортированных по частоте использования
      produces:
      - application/json
      responses:
        "200":
          description: Список наиболее используемых категорий
          schema:
            $ref: '#/definitions/dto.CategoriesListResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение наиболее используемых категорий
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"